refactor: move storeHost related code from BrokerController to BrokerNettyServer
